refactor: remove shared dependency from provider failures
